requirements.txt
smbus
pimoroni-bme280
picamera
pillow
Jinja2
Adafruit_DHT